+++ extra/github.com/sirupsen/logrus/otellogrus/example/main.go
@@ -1,0 +1,61 @@
+// Copyright The OpenTelemetry Authors
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//     http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
+package main
+
+import (
+	"context"
+	"log"
+
+	"github.com/sirupsen/logrus"
+
+	"go.opentelemetry.io/contrib/extra/github.com/sirupsen/logrus/otellogrus"
+	"go.opentelemetry.io/otel/api/global"
+	"go.opentelemetry.io/otel/exporters/stdout"
+	sdktrace "go.opentelemetry.io/otel/sdk/trace"
+)
+
+func main() {
+	initTracer()
+
+	// Add OpenTelemetry logging hook.
+	logrus.AddHook(otellogrus.NewLoggingHook())
+
+	tracer := global.Tracer("example")
+	ctx := context.Background()
+
+	ctx, span := tracer.Start(ctx, "main")
+	defer span.End()
+
+	logrus.WithContext(ctx).Error("hello")
+}
+
+func initTracer() {
+	// Create stdout exporter to be able to retrieve
+	// the collected spans.
+	exporter, err := stdout.NewExporter(stdout.WithPrettyPrint())
+	if err != nil {
+		log.Fatal(err)
+	}
+
+	// For the demonstration, use sdktrace.AlwaysSample sampler to sample all traces.
+	// In a production application, use sdktrace.ProbabilitySampler with a desired probability.
+	tp := sdktrace.NewTracerProvider(sdktrace.WithConfig(sdktrace.Config{DefaultSampler: sdktrace.AlwaysSample()}),
+		sdktrace.WithSyncer(exporter),
+	)
+	if err != nil {
+		log.Fatal(err)
+	}
+	global.SetTracerProvider(tp)
+}